Added possibility to switch off auth on redis Fixes https://github.com/testcontainers/testcontainers-spring-boot/issues/167
