+++ embedded-redis/src/test/java/com/playtika/test/redis/clustered/ClusterNoAuthEmbeddedRedisTest.java
@@ -21,22 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.playtika.test.redis;
+package com.playtika.test.redis.clustered;
 
-import com.playtika.test.common.properties.CommonContainerProperties;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.junit.Ignore;
+import org.springframework.test.context.ActiveProfiles;
 
-@Data
-@EqualsAndHashCode(callSuper = true)
-@ConfigurationProperties("embedded.redis")
-public class RedisProperties extends CommonContainerProperties {
-    public static final String BEAN_NAME_EMBEDDED_REDIS = "embeddedRedis";
-    public String dockerImage = "redis:4.0.12-alpine";
-    public String user = "root";
-    public String password = "passw";
-    public String host = "localhost";
-    public int port = 6379;
-    public boolean requirepass = true;
+@Ignore("Move from static port usage to dynamic one for cluster env")
+@ActiveProfiles("noauth")
+public class ClusterNoAuthEmbeddedRedisTest extends BaseClusterEmbeddedRedisTest {
+
 }
