+++ embedded-redis/src/test/resources/application-auth.properties
@@ -1,0 +1,2 @@
+spring.redis.client-name=${embedded.redis.user}
+spring.redis.password=${embedded.redis.password}
