+++ embedded-redis/src/test/resources/application-enabled.properties
@@ -1,0 +1,5 @@
+spring.redis.host=${embedded.redis.host}
+spring.redis.port=${embedded.redis.port}
+
+spring.redis.timeout=2500
+
