--- embedded-redis/src/test/resources/bootstrap-enabled.properties
@@ -1,1 +1,0 @@
-spring.redis.url=redis://${embedded.redis.user}:${embedded.redis.password}@${embedded.redis.host}:${embedded.redis.port}
