+++ embedded-redis/src/test/resources/bootstrap-noauth.properties
@@ -1,0 +1,1 @@
+embedded.redis.requirepass=false
